--- login-app/src/context/task-context.js
@@ -1,5 +1,0 @@
-import React from "react";
-
-const TaskContext = React.createContext({});
-
-export default TaskContext;
