+++ login-app/src/index.jsx
@@ -5,12 +5,15 @@ import "./index.css";
 import App from "./App";
 import ThemeContext from "./context/ThemeContext";
 import { AuthContextProvider } from "./context/auth-context";
+import { TaskContextProvider } from "./context/task-context";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 root.render(
   <ThemeContext>
     <AuthContextProvider>
-      <App />
+      <TaskContextProvider>
+        <App />
+      </TaskContextProvider>
     </AuthContextProvider>
   </ThemeContext>
 );
